refactor(challenge-v2): split modules by responsibility, rename for game/bond symmetry
