+++ crates/proof/challenge-v2/src/game_discovery.rs
